Move Route Advertisement Latency to measurements package
